Separate UpdateActivityTimer
